test: Fix role names in OpenShift marmalade test project
It is very confusing to pre-create the "amanda" and other users in our
OpenShift test VM:

  # oc get rolebindings -n marmalade
  NAME      ROLE      USERS                  GROUPS    SERVICE ACCOUNTS   SUBJECTS
  admin     /admin    scruffy
  edit      /edit     scruffy, amanda
  view      /view     scruffy, tom, amanda   sports

But these two roles aren't actually valid in OpenShift:

  # oadm policy add-role-to-user edit amanda -n marmalade
  The RoleBinding "edit" is invalid.

This causes these users/groups to be shown on the "Show all projects"
Users list and shown as being members of marmalade, but are not visible
on the marmalade project page (after actually creating the user by
logging in as them).

Adjust the "oc patch" command to create these roles with the official
registry-{editor,viewer} roles so that they actually work. This avoids
confusion for people who use our VM to test the cockpit registry.
Notably none of our tests actually rely on these two users/roles, these
are just for human testing.