test: Stop reloading sshd when Debian networking changes
Whenever NetworkManager or anything changes the network, Debian and
Ubuntu reloads sshd with SIGHUP. This leads to a real connectivity
issue.

In our tests this leads to intermittent connection failures. In real
life this would result in an admin or Ansible unable to reach the
system temporarily.

We choose to work around this problem for now by disabling the
broken behavior. In reality the best fix for this problem would be
to have sshd use socket activation on Debian and Ubuntu, the way
it does on Fedora

Closes #6158