test: Rename debian-8 to debian-stable
We always want to test on the latest stable and testing, so stop using
release numbers and use the "stable" alias instead. Explicitly change
the apt source in the bootstrap script, to be completely symmetric to
debian-testing.bootstrap and to ensure that we actually dist-upgrade to
a newer stable if it gets released, but virt-builder doesn't yet have
images for it yet.