bots: Move fedora-i386 image to Fedora 27
Fedora 25 is EOL now. Since Fedora 26 the mirrors don't publish i386
repositories any more, just a bi-arch x86_64/ repository. As this is not
straightforward to use with virt-install, move to virt-builder instead;
that still offers an i386 image and is faster, too.

Port fedora-25.setup's `HAVE_KUBERNETES` and conditional cockpit/base
docker build logic to fedora-27.setup to share it between 32 and 64 bit
images. Install docker explicitly as on 32 bit it won't get pulled in by
kubernetes.

Drop obsolete test skip in check-packagekit which now works on a Fedora
27 based image.