image: Add Kubevirt to Openshift image
Kubevirt (virtual machines in k8s) is added to the openshift test
image as an optional component.

For latest stable Kubevirt release, corresponding manifests are
stored under /kubevirt directory and related docker images are
pre-pulled.

Kubevirt is still not installed by default.

Following is needed to run virtual machines in the cluster:
- enable nested virtualization via `/root/nested-kvm`
- create objects via manifests, i.e.:
  for manifest in /kubevirt/*.yaml ; do oc create -f $manifest ; done
- create a VM, i.e.:
  oc create -f ${KUBEVIRT_HOME}/cluster/vm.yaml
- verify:
  oc get vm