test: Move cloud-init and SSH keys to bots/
These files are conceptually related to the machinery that builds and
provides images in bots/, not to the actual tests. This reduces the
number of symlinks in test/common/ and thus makes it easier for
third-party Cockpit modules to reuse the test infrastructure.

This temporarily introduces ugly relative paths from test/common to
bots/machine in testvm.py. This will be cleaned up in a subsequent
commit.

This is compatible with back-branches, as the keys and cloud-init.iso
did not change since the rhel-7.4 branch.