test: Move testvm.py to bots/
testvm.py provides the integration test API to the images in bots/, and
thus should be reusable for third-party projects that provide Cockpit
pages or apps. To make this possible with just checking out cockpit's
bots/ directory, without wrangling symlinks and replacing them with
files from other cockpit directories, move it to bots/machine/ and
change the tests to import it from there.

Use test/verify/parent.py from avocado and container tests as well, to
provide a single place that defines the Python module paths.

This should stay compatible with backbranches: make-checkout always
replaces the former bots/machine/testvm.py symlink with a copy of
master's version, but this is now the actual and current file (and thus
does not change behaviour). The integration tests themselves continue to
use test/common/testvm.py in back branches. So after landing this, the
symlink mangling code in make-checkout can be dropped.