bots: Add RHEL 8 nightly image
In rhel.setup:
 - Enable nightly RHEL 8 BaseOS and AppStream repositories for the
   "rhel-8" image.
 - Invert the storaged vs. udisks2 condition, to avoid having to extend
   the affected OS list in the future.
 - Don't install some packages which are not yet available in current
   RHEL 8 nightlies (libvirt, kubernetes, and udisks2 plugins).

Refresh the image every week to stay on top of current development, but
don't run this for every PR for the time being (i. e. similar to
continuous-atomic or fedora-testing).

As we don't yet have EPEL for RHEL-8 and thus no "mock" package, use the
fedora-27 image to build packages for rhel-8 for the time being. This
should be reasonably close right now, as RHEL 8 currently auto-imports
from Fedora 28.