bots: Drop obsolete rpcbind hack in openshift.setup
The corresponding bug got fixed in Fedora 25 already and does not affect
current RHEL releases.

Also temporarily disable the `set -x` for the version detection to avoid
the huge JSON spill in regular log files. It will still be shown on
failure anyway (but then just once and not three times).