bots: Update selenium hub and chrome to latest 3.x version
2.x is two years old, testing with these old browsers isn't really
useful any more. Upgrade hub and chrome to the latest 3.x version, which
can now  use the "3" tag.

Unfortunately node-firefox:3 does not work with our tests, it times out
with "Unable to attach remote Browser on hub". See
https://github.com/cockpit-project/cockpit/pull/9178#issuecomment-389778682
for some initial investigation. So keep firefox at the old 2.x for now.

The chrome browser name changed in the Grid Console (`googlechrome` →
`chrome`). To avoid false positives, check for the `browserName:` prefix
in the JSON dump.

Closes #9178