storage: Product version of NBDE
This rewrites most of cockpit-clevis-tool.py in JavaScript.  We still
need a small monitor hack until UDisks2 has enough API.

Support for plain passphrase slots has been added.

Support for editing existing tang configurations and passphrases has
been added.

Support for the "http" pin has been removed. It is too generic and
will be replaced with a more specific UI for FreeIPA Vault, say.

The "Check" button has been removed to reduce the scope of this
change.

LUKS2 formatted volumes are recognized and handled by omitting the UI
completely since Clevis can not yet work with LUKS2 (and we don't want
to have a version of the UI that doesn't need Clevis.)

Closes #9465