tools: Move cockpit-ssh from -dashboard to -bridge
We are not yet ready to install cockpit-dashboard by default everywhere
just yet, because of the insufficient isolation of different server
iframes in the browser.

But we always want to provide the "connect to remote server"
functionality through the login page, bastion host containers, or direct
URLs (when integrating Cockpit into e. g. Foreman). So ship the
cockpit-ssh bridge and the corresponding pkg/ssh/ manifest in
cockpit-bridge. This leaves cockpit-dashboard to ship only
pkg/dashboard, and thus it becomes an architecture independent package.

Adjust the package dependencies/conflicts accordingly. It is tempting to
bump the "requires/cockpit" in dashboard's manifest, but that would get
in the way of downstream RHEL/CentOS packaging, so just put the version
number into the spec/control files directly.

Also adjust the container build scripts, they need cockpit-ssh (but
without the actual dashboard bits). As this is now provided by
cockpit-bridge, install its dependency polkit-libs into the cockpit/base
test container.

Closes #9395