bots: Create proper mock on rhel-x image
Stop building rhel-x packages on fedora-28. There is an ever-increasing
set of differences which we can't represent this way, and we were not
testing what really was built and shipped downstream.

As there is no EPEL yet for rhel-x, grab the mock package from Fedora 28
(it's noarch and works fine), and supply a mock config for the Go
enabled buildroot.